FE/Qt4: suppress typecast warnings.
svn:sync-xref-src-repo-rev: r74093

+++ src/VBox/Frontends/VirtualBox/src/runtime/UIMouseHandler.cpp
@@ -1,4 +1,4 @@
-/* $Id: UIMouseHandler.cpp 38311 2011-08-04 13:08:39Z noreply@oracle.com $ */
+/* $Id: UIMouseHandler.cpp 38798 2011-09-20 09:41:10Z ramshankar.venkataraman@oracle.com $ */
 /** @file
  *
  * VBox frontends: Qt GUI ("VirtualBox"):
@@ -812,8 +812,8 @@ bool UIMouseHandler::mouseEvent(int iEventType, ulong uScreenId,
             double xRatio = scaledSize.isValid() ? (double)pFrameBuffer->width() / (double)scaledSize.width() : 1;
             double yRatio = scaledSize.isValid() ? (double)pFrameBuffer->height() / (double)scaledSize.height() : 1;
             /* Set scaling if scale-factor is present: */
-            cpnt.setX(cpnt.x() * xRatio);
-            cpnt.setY(cpnt.y() * yRatio);
+            cpnt.setX((int)(cpnt.x() * xRatio));
+            cpnt.setY((int)(cpnt.y() * yRatio));
 
             /* Bound coordinates: */
             if (cpnt.x() < 0) cpnt.setX(0);
